消息中心增加mq方式的数据接收处理；增加操作日志的接入；通道模板页面改造并优化相关代码

+++ message/src/main/java/com/diligrp/message/controller/MarketChannelController.java
@@ -3,7 +3,13 @@ package com.diligrp.message.controller;
 import cn.hutool.core.codec.Base64;
 import cn.hutool.json.JSONObject;
 import cn.hutool.json.JSONUtil;
+import com.dili.logger.sdk.annotation.BusinessLogger;
+import com.dili.logger.sdk.util.LoggerUtil;
 import com.dili.ss.domain.BaseOutput;
+import com.dili.uap.sdk.domain.UserTicket;
+import com.dili.uap.sdk.session.SessionContext;
+import com.diligrp.message.common.enums.MessageEnum;
+import com.diligrp.message.constants.MessageConstant;
 import com.diligrp.message.domain.MarketChannel;
 import com.diligrp.message.domain.TriggersTemplate;
 import com.diligrp.message.domain.vo.MarketChannelVo;
@@ -80,13 +86,23 @@ public class MarketChannelController {
      * @throws Exception
      */
     @RequestMapping(value="/save.action", method = {RequestMethod.GET, RequestMethod.POST})
-    public @ResponseBody BaseOutput insert(@RequestBody MarketChannel marketChannel) throws Exception{
+    @ResponseBody
+    @BusinessLogger(systemCode = MessageConstant.SYSTEM_CODE, businessType = "message_market_channel")
+    public BaseOutput save(@RequestBody MarketChannel marketChannel) throws Exception {
         marketChannel.setSecret(Base64.encode(marketChannel.getSecret()));
         marketChannel.setModified(LocalDateTime.now());
-        if (Objects.isNull(marketChannel.getId())){
+        String notes = "修改市场通道信息";
+        String prefix = "修改后为:";
+        String operationType = "edit";
+        if (Objects.isNull(marketChannel.getId())) {
             marketChannel.setCreated(LocalDateTime.now());
+            notes = "新增市场通道信息";
+            prefix = "";
+            operationType = "add";
         }
         marketChannelService.saveOrUpdate(marketChannel);
+        UserTicket userTicket = getUserTicket();
+        LoggerUtil.buildBusinessLoggerContext(marketChannel.getId(), String.valueOf(marketChannel.getId()), userTicket.getId(), userTicket.getRealName(), userTicket.getFirmId(), buildLoggerContent(marketChannel, prefix).toString(), notes, operationType);
         return BaseOutput.success("保存成功");
     }
 
@@ -116,15 +132,43 @@ public class MarketChannelController {
      */
     @RequestMapping(value="/delete.action", method = {RequestMethod.GET, RequestMethod.POST})
     @ResponseBody
+    @BusinessLogger(systemCode = MessageConstant.SYSTEM_CODE, businessType = "message_market_channel", notes = "删除市场通道信息", operationType = "del")
     public BaseOutput delete(Long id) {
-        if (Objects.isNull(id)){
+        if (Objects.isNull(id)) {
             return BaseOutput.failure("删除失败！参数ID为空");
         }
         List<TriggersTemplate> list = triggersTemplateService.listByMarketChannelId(String.valueOf(id));
-        if (CollectionUtils.isNotEmpty(list)){
+        if (CollectionUtils.isNotEmpty(list)) {
             return BaseOutput.failure("删除失败！该通道下还存在消息模板，不能删除！");
         }
-        marketChannelService.delete(id);
+        MarketChannel marketChannel = marketChannelService.get(id);
+        if (Objects.nonNull(marketChannel)) {
+            UserTicket userTicket = getUserTicket();
+            LoggerUtil.buildBusinessLoggerContext(marketChannel.getId(), String.valueOf(marketChannel.getId()), userTicket.getId(), userTicket.getRealName(), userTicket.getFirmId(), buildLoggerContent(marketChannel, "").toString());
+            marketChannelService.delete(id);
+        }
         return BaseOutput.success("删除成功");
+    }
+
+    /**
+     * 构建日志数据
+     * @param marketChannel 市场通道
+     * @return
+     */
+    private StringBuffer buildLoggerContent(MarketChannel marketChannel, String prefix) {
+        StringBuffer stringBuffer = new StringBuffer(prefix);
+        stringBuffer.append(" 市场: ").append(marketRpcService.getByCode(marketChannel.getMarketCode()).get().getName());
+        stringBuffer.append(" 通道: ").append(MessageEnum.ChannelEnum.getInstance(marketChannel.getChannel()).getName());
+        stringBuffer.append(" 签名: ").append(marketChannel.getSignature());
+        stringBuffer.append(" 账号: ").append(marketChannel.getAccessKey());
+        return stringBuffer;
+    }
+
+    /**
+     * 获取当前登录用户
+     * @return
+     */
+    private UserTicket getUserTicket() {
+        return SessionContext.getSessionContext().getUserTicket();
     }
 }
